requirements.txt
streamlit>=1.40.0
sentence-transformers>=3.0.0
psycopg2-binary>=2.9.0
pgvector>=0.3.0
groq>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.26.0
openpyxl>=3.1.0
